requirements.txt
Flask
groq
Flask-Cors
python-dotenv